+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 Key: [+] added, [-] removed, [~] modified, [!] important.
 
-## 18-04-2019_1
+## 19-04-2019_1
 + [+] Added lyrics command to enable users to lookup lyrics.
++ [+] Added urban dictionary command because it's both free and easy.
 
 ## 10-04-2019_1
 + [+] Added new database connection class for the automatic provisioning of shards.